KB-11836 | DEV | BE | Enablement for MDO Report via new hierarchy
1. Fixed the code review comments.
2. Implemented Cache for 4hours.
