+++ src/main/java/org/sunbird/orghierarchyreport/repository/MdoChildrenLookUpRepository.java
@@ -8,6 +8,6 @@ import java.util.List;
 
 public interface MdoChildrenLookUpRepository extends JpaRepository<MdoChildrenLookupEntity, String> {
 
-    @Query(value = "SELECT children_id from mdo_children_lookup where mdo_id in (?1)", nativeQuery = true)
-    List<String> findAllChildrenByMdoId(List<String> sbOrgIdList);
+    @Query(value = "SELECT * from mdo_children_lookup where mdo_id in (?1)", nativeQuery = true)
+    List<MdoChildrenLookupEntity> findAllChildrenByMdoId(List<String> sbOrgIdList);
 }
