+++ src/main/java/org/sunbird/orghierarchyreport/service/OrgLevelHierarchyService.java
@@ -4,6 +4,6 @@ import org.sunbird.common.model.SBApiResponse;
 
 import java.util.Map;
 
-public interface OrgHierarchyReportService {
+public interface OrgLevelHierarchyService {
     SBApiResponse orgExtSearchV3(Map<String, Object> request);
 }
