+++ src/main/resources/application.properties
@@ -610,3 +610,7 @@ kafka.topics.org.user.bulk.transfer.event.group=user-bulk-transfer-group
 org.hierarchy.bulk.user.transfer.container.name=userbulktransfer
 kafka.topics.org.user.bulk.transfer.event=dev.org.user.bulk.transfer
 org.hierarchy.identifier.regex.pattern=<([^>]+)>
+
+org.level.hierarchy.cache.key.ttl=14400
+org.level.hierarchy.es.limit=0
+org.level.hierarchy.es.offset=2000
